implemented and configured redux
