+++ package.json
@@ -22,9 +22,13 @@
     "react-native-message-bar": "^2.0.10",
     "react-native-responsive-screen": "^1.1.10",
     "react-native-simple-message-bar": "^1.0.3",
-    "react-navigation": "^2.16.0"
+    "react-navigation": "^2.16.0",
+    "react-redux": "^6.0.1",
+    "react-thunk": "^1.0.0",
+    "redux": "^4.0.1",
+    "redux-thunk": "^2.3.0"
   },
   "devDependencies": {
-    "jest-expo": "30.0.0"
+    "jest-expo": "^32.0.0"
   }
 }
