+++ src/App.tsx
@@ -1,0 +1,30 @@
+import React from 'react'
+import { Platform, StatusBar, StyleSheet, View } from 'react-native'
+import { Provider } from 'react-redux'
+import configureStore from './store/configureStore'
+import AppNavigator from './navigation/AppNavigator'
+
+const store = configureStore()
+
+interface Props {}
+
+export default class App extends React.Component<Props> {
+
+  render() {
+      return (
+        <Provider store={store}>
+          <View style={styles.container}>
+            {Platform.OS === 'ios' && <StatusBar barStyle="default" />}
+            <AppNavigator />
+          </View>
+        </Provider>
+      );
+    }
+  }
+
+  const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+  },
+});
