+++ src/store/reducers/index.ts
@@ -1,0 +1,7 @@
+import { combineReducers } from 'redux'
+
+const rootReducer = combineReducers({
+    // Here will put my reducers
+})
+
+export default rootReducer
